refactor(liveresultsview): changes according CR
