+++ libs/shared/src/components/diagram/bar-diagram/config.ts
