+++ libs/shared/src/components/diagram/bar-diagram/styled.ts
