+++ libs/shared/src/components/diagram/index.tsx
@@ -3,13 +3,12 @@ import Chart from "chart.js/auto";
 import ChartDataLabels from "chartjs-plugin-datalabels";
 import React from "react";
 
-import { DIAGRAM_CONFIG } from "./config";
-import * as Styled from "./styled";
+import { BarDiagram } from "./bar-diagram";
 
 Chart.register(LinearScale, ChartDataLabels);
 
 interface DiagramProps {
-  labels: string[] | undefined;
+  labels?: string[];
   data: ChartDataset<any, unknown>[];
   title: string;
   type: DiagramType;
@@ -29,57 +28,9 @@ export const Diagram: React.FC<DiagramProps> = ({ labels, data, title, type }) =
     ]
   };
 
-  const diagramOptions: Record<DiagramType, JSX.Element> = {
-    [DiagramType.Bar]: (
-      <Styled.DiagramBar
-        data={diagramData}
-        options={{
-          plugins: {
-            datalabels: {
-              anchor: "start",
-              align: "top",
-              offset: DIAGRAM_CONFIG.offset
-            },
-            title: {
-              display: true,
-              align: "start",
-              font: { size: DIAGRAM_CONFIG.fontSize },
-              padding: { top: DIAGRAM_CONFIG.paddingTop, bottom: DIAGRAM_CONFIG.paddingBottom },
-              text: title
-            },
-            legend: {
-              display: false
-            }
-          },
-          datasets: {
-            bar: {
-              borderRadius: DIAGRAM_CONFIG.borderRadius,
-              backgroundColor: DIAGRAM_CONFIG.bacgroundColorLight
-            }
-          },
-          scales: {
-            y: {
-              display: false,
-              grid: {
-                display: false
-              }
-            },
-            x: {
-              grid: {
-                display: false,
-                borderWidth: DIAGRAM_CONFIG.borderWidth,
-                borderColor: DIAGRAM_CONFIG.backgroundColor
-              },
-              ticks: {
-                showLabelBackdrop: true
-              }
-            }
-          }
-        }}
-        height={DIAGRAM_CONFIG.height}
-      />
-    )
+  const diagrams: Record<DiagramType, JSX.Element> = {
+    [DiagramType.Bar]: <BarDiagram diagramData={diagramData} title={title} />
   };
 
-  return diagramOptions[type];
+  return diagrams[type];
 };
