+++ libs/shared/src/data/feedback/index.ts
@@ -1,1 +1,2 @@
 export * from "./mocks";
+export * from "./types";
